+++ FlaneurOpen/Classes/Form Tools/FlaneurFormImagePickerElementCollectionViewCell.swift
@@ -36,14 +36,20 @@ class FlaneurFormImagePickerElementCollectionViewCell: FlaneurFormElementCollect
                            attribute: .bottom,
                            multiplier: 1.0,
                            constant: 4.0).isActive = true
-
         NSLayoutConstraint(item: launcherButton,
                            attribute: .height,
                            relatedBy: .equal,
-                           toItem: launcherButton,
-                           attribute: .width,
+                           toItem: nil,
+                           attribute: .notAnAttribute,
                            multiplier: 1.0,
-                           constant: 0.0).isActive = true
+                           constant: 80.0).isActive = true
+        NSLayoutConstraint(item: launcherButton,
+                           attribute: .width,
+                           relatedBy: .equal,
+                           toItem: nil,
+                           attribute: .notAnAttribute,
+                           multiplier: 1.0,
+                           constant: 80.0).isActive = true
     }
 
     override func configureWith(formElement: FlaneurFormElement) {