+++ CHANGELOG.md
@@ -23,6 +23,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * `FlaneurFilterCollectionViewCell` has been made `public` to allow UIAppearance customization #12
 * `FlaneurMapView` replaced `FlaneurMapViewController` #13
 * Updated project and dependencies for compatibility with Xcode 9
+* Right actions in a `FlaneurNavigationBar` can either be a collection if images or a single text #16
+* Right actions in a `FlaneurNavigationBar` are now have public getters for their `UIButton` #16
 
 ## [0.3.0] - 2017-09-19